backend/requirements.txt
tensorflow
datasets
flask
flask-cors
numpy
pillow

